% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.code.R
\name{default.code}
\alias{default.code}
\title{Set a default external simulation code}
\usage{
default.code(code = NULL)
}
\arguments{
\item{code}{structured list specifying the default external simulation code used when calling \code{\link{run.simulation}}. This list has exactly the same format as the sub-list `code` described in the documentation of \code{\link{run.simulation}}.}
}
\value{
Returns the current list `code`. If no such last has been set and `default.code()` is called without argument, an error is produced.
}
\description{
Set a default external simulation code
}
\author{
Danail Obreschkow
}
