% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbody.env.R
\docType{data}
\name{.nbody.env}
\alias{.nbody.env}
\title{Package environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.nbody.env
}
\description{
Environment used to store paths and options for external code.
}
\seealso{
\code{\link{default.code}}
}
\author{
Danail Obreschkow
}
\keyword{datasets}
