% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ics.R
\name{ics.earth}
\alias{ics.earth}
\title{Make ICs for N-body simulation}
\usage{
ics.earth()
}
\arguments{
\item{sim}{is an output of \code{run.simulation}}
}
\description{
Description
}
\author{
Danail Obreschkow
}
