% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbody-package.R
\docType{package}
\name{nbody}
\alias{nbody}
\alias{nbody-package}
\title{Gravitational N-body simulator}
\description{
Gravitational N-body solver for teaching and entry-level scientific computing.
}
\author{
Danail Obreschkow
}
