% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot.simulation}
\alias{plot.simulation}
\title{Visualize an N-body simulation}
\usage{
\method{plot}{simulation}(sim, units, cex = 0.3, pch = 20,
  title = "", ...)
}
\arguments{
\item{sim}{is an output of \code{run.simulation}}
}
\description{
Description
}
\author{
Danail Obreschkow
}
