% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\alias{setup.earth}
\alias{setup.halley}
\alias{setup.ellipse}
\alias{setup.sunearth}
\title{Initialize N-body simulation}
\usage{
setup()

setup.sunearth(t.max = cst$yr, dt.out = cst$yr, ...)

setup.halley(
  t.max = NULL,
  nperiods = 1,
  dt.out = 3 * cst$month,
  e = 0.96714,
  s = 17.834 * cst$AU,
  ...
)

setup.ellipse(t.max = NULL, nperiods = 1, e = 0.9, s = 1, f = 0.5, ...)
}
\arguments{
\item{t.max}{final simulation time}

\item{dt.out}{time step for simulation output}

\item{...}{other simulation parameters used by \code{\link{run.simulation}}}

\item{nperiods}{number of orbital periods to be computed; ignored if \code{t.max} is specified.}

\item{e}{eccentricity}

\item{s}{semi-major axis}

\item{f}{mass-ratio}
}
\value{
default routine, which is identical to setup.halley()

\code{setup.sunearth()} sets up a simple 2-body simulation of the Earth around the Sun, using only approximate orbital specifications.

\code{setup.halley()} sets up a 2-body simulation of Halley's Commet around the Sun.

\code{setup.ellipse()} sets up an elliptical Keplerian orbit in natural units
}
\description{
Routines to generate the structured lists of initial conditions and simulation parameters required to run an N-body simulation with \code{\link{run.simulation}}.
}
\examples{
sim = setup.halley()
sim = run.simulation(sim)
plot(sim)

}
\author{
Danail Obreschkow
}
