% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cst.R
\docType{data}
\name{cst}
\alias{cst}
\title{Physical constants}
\format{An object of class \code{list} of length 17.}
\usage{
cst
}
\description{
Useful physical constants in SI units.
}
\details{
Some of these constants are used internally by routines of the \code{nbody} package. Note that even if these constants are modified (e.g. via \code{cst$G=1}), the routines of the \code{nbody} package will still access the original values.
}
\examples{
summary(cst)

}
\author{
Danail Obreschkow
}
\keyword{datasets}
