% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.simulation.R
\name{run.simulation}
\alias{run.simulation}
\title{Run a direct N-body simulation}
\usage{
run.simulation(sim, measure.time = TRUE)
}
\arguments{
\item{sim}{structured list of simulation settings, which must contain the following sublists:\cr\cr

\code{ics} is the sublist of initial conditions. It must contain the items: \cr
\code{m} = N-vector with the masses of the N particles\cr
\code{x} = N-by-3 matrix specifying the initial position in cartesian coordinates\cr
\code{v} = N-by-3 matrix specifying the initial velocities\cr\cr

\code{para} is the sublist of simulation parameters. It contains the items:\cr
\code{t.max} = final simulation time.\cr
\code{dt.xmax} = maximum time step; ignored if set to \code{NULL} (default).\cr
\code{dt.out} = output time step, i.e. time step between successive snapshots in the \code{output} sublist returned by \code{run.simulation}.\cr
\code{eta} = scaling of adaptive time step. Smaller values lead to proportionally smaller adaptive time steps. Typical values range between 0.001 and 0.1. The default is 0.01.\cr
\code{algorithm} = character string specifying the integrator to be used. Currently implemented algorithms are 'euler' (1st order), 'leapfrog' (2nd order, default), 'yoshida' (4th order).\cr
\code{rsmooth} = optional smoothing radius. No smoothing is assumed by default.\cr
\code{afield} = a function(x,t) of position \code{x} (3-vector) and time \code{t} (scalar), specifying the external acceleration field. No field is assumes if set to \code{NULL} (default).}

\item{measure.time}{logical flag that determines whether time computation time will be measured and displayed.}
}
\value{
The routine returns the structured list of the input argument, with one sublist \code{output} added. This sublist contains the items:
\item{t}{k-vector with the simulation times of the k output snapshots.}
\item{x}{k-by-N-by-3 array giving the 3D coordinates of the N particles in k snapshots.}
\item{v}{k-by-N-by-3 array giving the 3D velocities of the N particles in k snapshots.}
\item{n.iterations}{Number of iterations used to run the simulation.}
}
\description{
Run direct N-body simulations using a common adaptive time step.
}
\details{
For a detailed description of the simulation method, please refer to the lecture notes on N-body simulations by Obreschkow (2019).
}
\examples{
sim = setup.halley()
sim = run.simulation(sim)
plot(sim, units=cst$AU, xlim=c(-20,60), ylim=c(-40,40), xlab='[AU]', ylab='[AU]')
cat(sprintf('This simulation was run with \%d iterations.\\n',sim$output$n.iterations))

}
\author{
Danail Obreschkow
}
\keyword{N-body}
\keyword{simulation}
