% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run.simulation}
\alias{run.simulation}
\title{Run a direct N-body simulation}
\usage{
run.simulation(sim, measure.time = TRUE)
}
\arguments{
\item{sim}{is a list of simulation settings:\cr\cr
\code{x} is a M-by-D array of M points, defining a regular cartesian grid in the D-dimensional observable space.\cr\cr
\code{xmin} and \code{xmax} are D-vectors specifying the lower and upper boundary of the grid in the D-dimensional observable space.\cr\cr
\code{dx} is a D-vector specifying the steps between grid points.\cr\cr}

\item{measure.time}{is a logical flag that determines whether time computation time will be measured and displayed.}
}
\value{
The routine returns a structured list \code{sim} with the following sublists:

\item{selection}{is a list describing the selection function of the data used when fitting the generative DF. The most important entries in this list are:\cr\cr
\code{veff} is a function of a D-dimensional vector, specifying the effective volume associated with an object of properties xval. If LSS is corrected for, i.e. if the argument \code{correct.lss.bias} is set to \code{TRUE}, this function is the final effecive volume, including the effect of LSS.\cr\cr
\code{veff.no.lss} is a function of a D-dimensional vector, specifying the effecive volume associate with an object, if LSS were not accounted for. This function is indentical to \code{veff}, if LSS is not accounted for in the fit, i.e. if the argument \code{correct.lss.bias} is set to \code{FALSE}.\cr\cr
\code{mode} is an integer specifying the format of the input argument \code{selection}.}

\item{fit}{is a list describing the fitted generative distribution function. Its most important entries are:\cr\cr
\code{p.best} is a P-vector giving the most likely model parameters according to the MML method.\cr\cr
\code{p.sigma} and \code{p.covariance} are the standard deviations and covariance matrices of the best-fitting parameters in the Gaussian approximation from the Hessian matrix of the modified likelihood function.\cr\cr
 \code{gdf} is a function of a D-dimensional vector, which is the generative DF, evaluated at the parameters \code{p.best}.\cr\cr
 \code{scd} is a function of a D-dimensional vector, which gives the predicted source counts of the most likely model, i.e. scd(x)=gdf(x)*veff(x).}
}
\description{
This function finds the most likely P-dimensional model parameters of a D-dimensional distribution function (DF) generating an observed set of N objects with D-dimensional observables x, accounting for measurement uncertainties and a user-defined selection function. For instance, if the objects are galaxies, \code{dffit} can fit a mass function (D=1), a mass-size distribution (D=2) or the mass-spin-morphology distribution (D=3). A full description of the algorithm can be found in Obreschkow et al. (2017).
}
\details{
For a detailed description of the method, please refer to the peer-reviewed publication by Obreschkow et al. 2017 (in prep.).
}
\examples{
sim = list(ics = ics.halley(), para = list(t.max = 76*cst$yr, dt.out = 3*cst$month))
sim = run.simulation(sim)
plot.simulation(sim, cst$AU, xlim=c(-20,60), ylim=c(-40,40), xlab='[AU]', ylab='[AU]')

}
\author{
Danail Obreschkow
}
\keyword{N-body}
\keyword{simulation}
